refactor: bookTitle과 seriesNumber를 headerView로 분리
